LPS-43689 Move classes to proper package
